// src/types/user.interface.ts
import { UserRole } from "@/lib/auth-utils";
import { IAdmin } from "./admin.interface";

export interface UserInfo {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  needPasswordChange: boolean;
  status: "ACTIVE" | "BLOCKED" | "DELETED";
  admin?: IAdmin;

  createdAt: string;
  updatedAt: string;
}